+++ changelog.md
@@ -13,6 +13,11 @@
 - `Fixed` 对bug的修复
 - `Security` 对安全的改进
 
+## [1.10.2] 2020-07-17
+
+- [Changed] `WeixinAuthProvider.signIn()` 新增 `syncUserInfo` 参数，同步微信账号信息作为用户信息
+- [Changed] `WeixinAuthProvider.getRedirectResult()` 新增 `syncUserInfo` 参数，同步微信账号信息作为用户信息
+
 ## [1.10.1] 2020-07-16
 - [Fix] 修复 `callFunction` 入参语法提示错误
 - [Fix] 优化数据库 API 错误处理